1. Labelset add/remove labels page: Add calcification rate availability info to the label-info popups. 2. Move labelset import tests to a different file from the other labelset-view tests.
