+++ project/labels/templates/labels/label_detail_box.html
@@ -1,6 +1,7 @@
 {% load static %}
 {% load truncate_float from common_tags %}
-{% load labels %}
+
+{% load calcify_rate_indicator from calcification %}
 
 
 <div class="label-basic-info-and-thumbnail-wrapper">
@@ -14,7 +15,7 @@
   {% if label.thumbnail %}
     {# Placeholder src to delay loading the actual image until it is shown. #}
     {# Actual image is in the data-src attribute. #}
-    {# Placeholder is 1x1 transparent, from: http://png-pixel.com/ #}
+    {# Placeholder is 1x1 transparent, from: https://png-pixel.com/ #}
     <img src="data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAAEAAAABCAQAAAC1HAwCAAAAC0lEQVR42mNkYAAAAAYAAjCB0C8AAAAASUVORK5CYII="
          data-src="{{ label.thumbnail.url }}"
          class="label-thumbnail lazy-load" alt="{{ label.name }}"/>
@@ -23,6 +24,10 @@
          class="label-thumbnail" alt="{{ label.name }}" />
   {% endif %}
 
+</div>
+
+<div class="line">
+  Calcification rate data: {% calcify_rate_indicator label 'text' %}
 </div>
 
 {% if label.description %}
