+++ project/labels/templates/labels/label_list.html
@@ -1,5 +1,8 @@
 {% extends "base.html" %}
-{% load calcify_icon popularity_bar status_icon from labels %}
+
+{% load calcify_rate_indicator from calcification %}
+{% load popularity_bar status_icon from labels %}
+
 
 {% block title %}Label list - CoralNet{% endblock %}
 
@@ -59,7 +62,7 @@
         <td>{% popularity_bar label %}</td>
         <td class="status-cell">
           {% status_icon label %}
-          {% calcify_icon label has_calcify_info %}
+          {% calcify_rate_indicator label 'icon' %}
         </td>
         <td>{{ label.default_code }}</td>
       </tr>
